% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnovaDesign.R
\name{AnovaDesign}
\alias{AnovaDesign}
\title{ANOVA tables of \pkg{DoETeachR} objects}
\usage{
AnovaDesign(formula, design, alpha.risk.model = 0.05)
}
\arguments{
\item{formula}{A formula that defines the model. Only order two effects are supported at maximum.}

\item{design}{A design from \code{\link[DoETeachR]{gen.FrFD}}, \code{\link[DoETeachR]{gen.FFD}} or \code{\link[DoETeachR]{gen.CCD}}.}

\item{alpha.risk.model}{The alpha risk of the global test of the model. If the global test shows a p.value larger than \bold{alpha.risk.model}, then all effects from the model are considered non significant.}
}
\value{
A list where the first element is the ANOVA table and the second is the adjusted R-squared of the model.
}
\description{
Computes the (type III) ANOVA table of a \pkg{DoETeachR} object using the model from \bold{formula} and the corresponding adjusted R-squared.
}
\examples{

design=gen.FFD(responses=c("R1"),factors=LETTERS[1:4])
design$R1=runif(nrow(design))
AnovaDesign(R1~.*.,design)
}
