% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AugmentDesign.R
\name{AugmentDesign}
\alias{AugmentDesign}
\title{Augment a design of experiments from \pkg{DoETeachR}}
\usage{
AugmentDesign(design, randomize = FALSE, star.res = 5, alpha = "rotatable")
}
\arguments{
\item{design}{A design from \code{\link[DoETeachR]{gen.FrFD}}, \code{\link[DoETeachR]{gen.FFD}} or \code{\link[DoETeachR]{gen.CCD}}.}

\item{randomize}{Logical, if TRUE, added runs are randomized.}

\item{star.res}{A positive integer giving the resolution at which \code{\link[DoETeachR]{AugmentDesign}} adds star points rather than a fraction for Fractional Factorial Designs.}

\item{alpha}{A character among \code{"rotatable"}, \code{"spherical"} or \code{"faced"} that specify the position of the star points.}
}
\value{
A data.frame where each row corresponds to a run, the first \code{length(factors)} columns are the factors and the last columns are the responses filled with \code{NA} for the new runs.
}
\description{
Augments a design of experiments from \pkg{DoETeachR} depending on its type. See details.
}
\details{
For FFD it adds star points. For FrFD it adds a fraction up to a \bold{star.res} resolution design and then adds star points. For CCD, it replicates the design.
}
\examples{
design=gen.FFD(responses=c("R1"),factors=LETTERS[1:4])
AugmentDesign(design)

design=gen.FrFD(responses=c("R1"),factors=LETTERS[1:4])
AugmentDesign(design)

design=gen.CCD(responses=c("R1"),factors=LETTERS[1:4])
AugmentDesign(design)
}
