% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.CCD.R
\name{gen.CCD}
\alias{gen.CCD}
\title{Generate Central Composite Design (CCD)}
\usage{
gen.CCD(
  responses,
  factors,
  ncenter = 3,
  randomize = FALSE,
  alpha = "rotatable"
)
}
\arguments{
\item{responses}{A character vector specifying the name(s) of the response(s).}

\item{factors}{A character vector specifying the names of the factors.}

\item{ncenter}{The number of center points to be added in the design.}

\item{randomize}{Logical, if TRUE, runs are randomized.}

\item{alpha}{A character among \code{"rotatable"}, \code{"spherical"} or \code{"faced"} that specify the position of the star points.}
}
\value{
A data.frame where each row corresponds to a run, the first \code{length(factors)} columns are the factors and the last columns are the responses filled with \code{NA}.
}
\description{
Generates a Central Composite Design (CCD) according to the specified parameters.
}
\examples{
gen.CCD(responses=c("R1"),factors=LETTERS[1:4])
}
