% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.FFD.R
\name{gen.FFD}
\alias{gen.FFD}
\title{Generate Full Factorial Design (FFD)}
\usage{
gen.FFD(responses, factors, ncenter = 0, randomize = FALSE)
}
\arguments{
\item{responses}{A character vector specifying the name(s) of the response(s).}

\item{factors}{A character vector specifying the names of the factors.}

\item{ncenter}{The number of center points to be added in the design.}

\item{randomize}{Logical, if TRUE, runs are randomized.}
}
\value{
A data.frame where each row corresponds to a run, the first \code{length(factors)} columns are the factors and the last columns are the responses filled with \code{NA}.
}
\description{
Generates a Full Factorial Design (FFD) according to the specified parameters.
}
\examples{
gen.FFD(responses=c("R1"),factors=LETTERS[1:4])
}
