% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.FrFD.R
\name{gen.FrFD}
\alias{gen.FrFD}
\title{Generate Fractional Factorial Design (FrFD)}
\usage{
gen.FrFD(responses, factors, frac = 2^1, ncenter = 0, randomize = FALSE)
}
\arguments{
\item{responses}{A character vector specifying the name(s) of the response(s).}

\item{factors}{A character vector specifying the names of the factors.}

\item{frac}{An integer power of 2. The fraction of the full factorial design. For example \code{frac=2} corresponds to half the full factorial design i.e. \code{2^(length(factors)-1)} runs.}

\item{ncenter}{The number of center points to be added in the design.}

\item{randomize}{Logical, if TRUE, runs are randomized.}
}
\value{
A data.frame where each row corresponds to a run, the first \code{length(factors)} columns are the factors and the last columns are the responses filled with \code{NA}.
}
\description{
Generates a Fractional Factorial Design (FrFD) according to the specified parameters.
}
\examples{
gen.FrFD(responses=c("R1"),factors=LETTERS[1:4])
}
