% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalPlot.R
\name{NormalPlot}
\alias{NormalPlot}
\title{Plot Normality of Effects of \pkg{DoETeachR} objects}
\usage{
NormalPlot(formula, design)
}
\arguments{
\item{formula}{A formula that defines the model. Only order two effects are supported at maximum.}

\item{design}{A design from \code{\link[DoETeachR]{gen.FrFD}} that has no degree of freedom for the residuals given the model of \bold{formula}}
}
\value{
A plot of normality of effects.
}
\description{
Plots the empirical cumulative probabilities against the absolute effects from \bold{formula} of a \pkg{DoETeachR} object. Adds a reference line following Zahn, D. (1975) method.
}
\examples{
design=gen.FrFD(responses=c("R1"),factors=LETTERS[1:4],frac=2^1)
design$R1=runif(nrow(design))
NormalPlot(R1~.*.,design)
}
\references{
Zahn, D (1975) Modifications of and Revised Critical Values for the Half-Normal Plot. Technometrics 17(2), 189-200
}
