% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimDesign.R
\name{OptimDesign}
\alias{OptimDesign}
\title{Compute Optimal Factor Values for a Response Target of \pkg{DoETeachR} Objects}
\usage{
OptimDesign(formula, design, bounds = NULL, target.optim = "min")
}
\arguments{
\item{formula}{A formula that defines the model. Only order two effects are supported at maximum.}

\item{design}{A design from \code{\link[DoETeachR]{gen.FrFD}}, \code{\link[DoETeachR]{gen.FFD}} or \code{\link[DoETeachR]{gen.CCD}}.}

\item{bounds}{Either NULL (Default) or a numeric vector of length two defining the range of values that must be considered for the factors. If NULL, the range of values for the factors are the minimum and maximum values of the plotted factors present in \bold{formula}.}

\item{target.optim}{Either \code{"min"}, \code{"max"} or a numeric value corresponding to a target value of the response indicating whether the objective is to minimize, maximize or maintain the value of the response at the target respectively.}
}
\value{
A list if length two where the first element is the optimal values of the factors and the second element is the value of the response obtained at the optimal values of the factors.
}
\description{
Computes the optimal values of the factors to reach \bold{target.optim} of the response given \bold{formula}.
}
\examples{
design=gen.CCD(responses=c("R1"),factors=LETTERS[1:4])
design$R1=runif(nrow(design))
OptimDesign(R1~.*.+I(A^2)+I(B^2)+I(C^2)+I(D^2),design)
}
