% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurfacePlot.R
\name{SurfacePlot}
\alias{SurfacePlot}
\title{Plot Response Surface of \pkg{DoETeachR} objects}
\usage{
SurfacePlot(
  formula,
  design,
  plotted.factors = NULL,
  bounds = NULL,
  target.optim = "min",
  value.non.plotted = "optimal"
)
}
\arguments{
\item{formula}{A formula that defines the model. Only order two effects are supported at maximum.}

\item{design}{A design from \code{\link[DoETeachR]{gen.FrFD}}, \code{\link[DoETeachR]{gen.FFD}} or \code{\link[DoETeachR]{gen.CCD}}.}

\item{plotted.factors}{Either NULL (Default) of character vector of length two naming the factors of \bold{formula} for which the response surface must be plotted. If NULL, the two plotted factors are those whose sum of squares of effects are the largest.}

\item{bounds}{Either NULL (Default) or a numeric vector of length two defining the range of values that must be considered for the \bold{plotted.factors}. If NULL, the range of values are the minimum and maximum values of the plotted factors present in \bold{formula}.}

\item{target.optim}{Either \code{"min"}, \code{"max"} or a numeric value corresponding to a target value of the response indicating whether the objective is to minimize, maximize or maintain the value of the response at the target respectively.}

\item{value.non.plotted}{Either \code{"optimal"} or \code{"central"} indicating whether the non plotted factors should be considered at their optimal value or their central value when drawing the response surface of \bold{plotted.factors}.}
}
\value{
A surface response plot
}
\description{
Plots the response surface from \bold{formula} as a function of \bold{plotted.factors} of a \pkg{DoETeachR} object. A green point is added at the optimal value depending on the target defined by \bold{target.optim}.
}
\examples{

design=gen.CCD(responses=c("R1"),factors=LETTERS[1:4])
design$R1=runif(nrow(design))
SurfacePlot(R1~.*.+I(A^2)+I(B^2)+I(C^2)+I(D^2),design)
}
