% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoETeachR-package.R
\docType{package}
\name{DoETeachR-package}
\alias{DoETeachR}
\alias{DoETeachR-package}
\title{DoETeachR: Design of Experiments for Higher Education}
\description{
A collection of functions for higher education of basic designs of experiments namely full factorial designs, fractional factorial designs and central composite designs.
}
\author{
\strong{Maintainer}: Benjamin Mahieu \email{benjamin.mahieu@oniris-nantes.fr}

Authors:
\itemize{
  \item Benjamin Mahieu \email{benjamin.mahieu@oniris-nantes.fr}
}

}
\keyword{internal}
